suse/redhat: optimize vagrant plugin vagrant-libvirt check
Unfortunately vagrant has no command line way to check if a plugin
is installed. Ansible doesn't like it to pipe things, and so there
is no easy way to check with ansible if a vagrant plugin is installed.

We want an optimal solution as otherwise we just try to install
vagrant-libvirt, and even if that is already installed this takes
*forever*. This is just unacceptable.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>